+++ src/sql/plot.py
@@ -52,7 +52,7 @@ FROM (
     SELECT "{{column}}"
     FROM "{{table}}"
     WHERE "{{column}}" <= {{hival}}
-)
+) AS _whishi
 """
     )
 
@@ -60,7 +60,7 @@ FROM (
 
     if with_:
         query = str(store.render(query, with_=with_))
-
+    query = sql.connection.Connection._transpile_query(query)
     values = con.execute(query).fetchone()
     keys = ["N", "wiskhi_max"]
     return {k: float(v) for k, v in zip(keys, values)}
@@ -74,7 +74,7 @@ FROM (
     SELECT "{{column}}"
     FROM "{{table}}"
     WHERE "{{column}}" >= {{loval}}
-)
+) AS _whislo
 """
     )
 
@@ -82,7 +82,7 @@ FROM (
 
     if with_:
         query = str(store.render(query, with_=with_))
-
+    query = sql.connection.Connection._transpile_query(query)
     values = con.execute(query).fetchone()
     keys = ["N", "wisklo_min"]
     return {k: float(v) for k, v in zip(keys, values)}
@@ -100,7 +100,7 @@ FROM "{{table}}"
 
     if with_:
         query = str(store.render(query, with_=with_))
-
+    query = sql.connection.Connection._transpile_query(query)
     values = con.execute(query).fetchone()[0]
     return values
 
@@ -118,7 +118,7 @@ OR  "{{column}}" > {{whishi}}
 
     if with_:
         query = str(store.render(query, with_=with_))
-
+    query = sql.connection.Connection._transpile_query(query)
     results = [float(n[0]) for n in con.execute(query).fetchall()]
     return results
 
@@ -280,7 +280,7 @@ FROM "{{table}}"
 
     if with_:
         query = str(store.render(query, with_=with_))
-
+    query = sql.connection.Connection._transpile_query(query)
     min_, max_ = con.execute(query).fetchone()
     return min_, max_
 
@@ -372,11 +372,13 @@ group by 1
 order by 1;
 """
     )
+
     query = template.render(table=table, column=column, bin_size=bin_size)
 
     if with_:
         query = str(store.render(query, with_=with_))
 
+    query = sql.connection.Connection._transpile_query(query)
     data = conn.execute(query).fetchall()
     bin_, height = zip(*data)