+++ src/tests/integration/test_generic_db_opeations.py
@@ -1,4 +1,5 @@
 import shutil
+from matplotlib import pyplot as plt
 import pytest
 import warnings
 from sql.telemetry import telemetry
@@ -35,8 +36,19 @@ def mock_log_api(monkeypatch):
 )
 def test_query_count(ip_with_dynamic_db, excepted, request):
     ip_with_dynamic_db = request.getfixturevalue(ip_with_dynamic_db)
-    out = ip_with_dynamic_db.run_line_magic("sql", "SELECT * FROM taxi LIMIT 3")
-    assert len(out) == excepted
+    out_normal_query = ip_with_dynamic_db.run_line_magic(
+        "sql", "SELECT * FROM taxi LIMIT 3"
+    )
+    assert len(out_normal_query) == excepted
+
+    # Test query with --with & --save
+    ip_with_dynamic_db.run_cell(
+        "%sql --save taxi_subset --no-execute SELECT * FROM taxi LIMIT 3"
+    )
+    out_query_with_save_arg = ip_with_dynamic_db.run_cell(
+        "%sql --with taxi_subset SELECT * FROM taxi_subset"
+    )
+    assert len(out_query_with_save_arg.result) == excepted
 
 
 # Create
@@ -145,6 +157,94 @@ def test_telemetry_execute_command_has_connection_info(
             },
         },
     )
+
+
+@pytest.mark.parametrize(
+    "cell",
+    [
+        ("%sqlplot histogram --table plot_something --column x"),
+        ("%sqlplot hist --table plot_something --column x"),
+        ("%sqlplot histogram --table plot_something --column x --bins 10"),
+    ],
+    ids=[
+        "histogram",
+        "hist",
+        "histogram-bins",
+    ],
+)
+@pytest.mark.parametrize(
+    "ip_with_dynamic_db",
+    [
+        ("ip_with_postgreSQL"),
+        ("ip_with_mySQL"),
+        ("ip_with_mariaDB"),
+        ("ip_with_SQLite"),
+        ("ip_with_duckDB"),
+    ],
+)
+def test_sqlplot_histogram(ip_with_dynamic_db, cell, request):
+    # clean current Axes
+    ip_with_dynamic_db = request.getfixturevalue(ip_with_dynamic_db)
+    plt.cla()
+
+    ip_with_dynamic_db.run_cell(
+        "%sql --save plot_something_subset"
+        " --no-execute SELECT * from plot_something LIMIT 3"
+    )
+    out = ip_with_dynamic_db.run_cell(cell)
+
+    assert type(out.result).__name__ in {"Axes", "AxesSubplot"}
+
+
+BOX_PLOT_FAIL_REASON = (
+    "Known issue, the SQL engine must support percentile_disc() SQL clause"
+)
+
+
+@pytest.mark.parametrize(
+    "cell",
+    [
+        "%sqlplot boxplot --table plot_something --column x",
+        "%sqlplot box --table plot_something --column x",
+        "%sqlplot boxplot --table plot_something --column x --orient h",
+        "%sqlplot boxplot --with plot_something_subset --table "
+        "plot_something_subset --column x",
+    ],
+    ids=[
+        "boxplot",
+        "box",
+        "boxplot-horizontal",
+        "boxplot-with",
+    ],
+)
+@pytest.mark.parametrize(
+    "ip_with_dynamic_db",
+    [
+        pytest.param("ip_with_postgreSQL"),
+        pytest.param(
+            "ip_with_mySQL", marks=pytest.mark.xfail(reason=BOX_PLOT_FAIL_REASON)
+        ),
+        pytest.param(
+            "ip_with_mariaDB", marks=pytest.mark.xfail(reason=BOX_PLOT_FAIL_REASON)
+        ),
+        pytest.param(
+            "ip_with_SQLite", marks=pytest.mark.xfail(reason=BOX_PLOT_FAIL_REASON)
+        ),
+        pytest.param("ip_with_duckDB"),
+    ],
+)
+def test_sqlplot_boxplot(ip_with_dynamic_db, cell, request):
+    # clean current Axes
+    ip_with_dynamic_db = request.getfixturevalue(ip_with_dynamic_db)
+    plt.cla()
+    ip_with_dynamic_db.run_cell(
+        "%sql --save plot_something_subset"
+        " --no-execute SELECT * from plot_something LIMIT 3"
+    )
+
+    out = ip_with_dynamic_db.run_cell(cell)
+
+    assert type(out.result).__name__ in {"Axes", "AxesSubplot"}
 
 
 @pytest.mark.parametrize(