+++ src/tests/test_telemetry.py
@@ -116,10 +116,11 @@ def test_sqlrender_telemetry_execution(mock_log_api, ip, simple_file_path_iris):
     # Simulate the sqlrender query
     ip.run_cell("%sql duckdb://")
     ip.run_cell(
-        "%sql --save class_setosa --no-execute \
-            SELECT * FROM read_csv_auto('"
+        "%sql --save class_setosa --no-execute "
+        "SELECT * FROM read_csv_auto('"
         + simple_file_path_iris
-        + "' WHERE class='Iris-setosa'"
+        + "')"
+        + " WHERE class='Iris-setosa'"
     )
     ip.run_cell("%sqlrender class_setosa")